pages/requirements.txt
deta==1.2.0
email_validator==1.3.1
keras==2.6.0
pandas==1.1.5
password_strength==0.0.3.post2
python-dotenv==1.0.0
streamlit==1.10.0
streamlit_option_menu==0.3.6
tensorflow==2.6.2
